requirements.txt
Flask
python-dotenv
requests
python-magic
PyPDF2
-e